requirements.txt
fastapi
uvicorn[standard]
python-pptx
python-dotenv
groq
